refactor: fakeViewModelを別ファイルに分割
