+++ app/src/main/java/jp/co/yumemi/droidtraining/WeatherMainViewModel.kt
@@ -4,12 +4,9 @@ import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
-import jp.co.yumemi.droidtraining.model.WeatherInfoData
-import jp.co.yumemi.droidtraining.repository.WeatherInfoDataRepository
 import jp.co.yumemi.droidtraining.usecases.GetWeatherInfoDataUseCase
 import jp.co.yumemi.droidtraining.usecases.UpdateWeatherInfoDataUseCase
 import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 import javax.inject.Inject
@@ -57,29 +54,4 @@ open class WeatherMainViewModel @Inject constructor(
     }
 }
 
-// 以下はプレビュー用のフェイク（天気情報初期値を設定できる）
-class FakeWeatherMainViewModel(
-    initialWeatherInfoData: WeatherInfoData,
-) : WeatherMainViewModel(
-    updateWeatherInfoDataUseCase = UpdateWeatherInfoDataUseCase(
-        FakeWeatherInfoDataRepository(initialWeatherInfoData),
-    ),
-    getWeatherInfoDataUseCase = GetWeatherInfoDataUseCase(
-        FakeWeatherInfoDataRepository(initialWeatherInfoData),
-    ),
-    savedStateHandle = SavedStateHandle(), // fake(empty)
-)
 
-class FakeWeatherInfoDataRepository(private val initialWeatherInfoData: WeatherInfoData) :
-    WeatherInfoDataRepository {
-    override val weatherInfoData: StateFlow<WeatherInfoData>
-        get() = MutableStateFlow(initialWeatherInfoData).asStateFlow()
-
-    override suspend fun updateWeatherInfoData() {
-        // do nothing
-    }
-
-    override fun setWeatherInfoData(weatherInfoData: WeatherInfoData) {
-        // do nothing
-    }
-}
